Refactor: 이미지 최적화 함수 분리
